review(tproxy): address codex findings — PATH hardening, input bounds, stale docs
From the codex review of the slice (no Critical/no relay bugs found):

- HIGH: mq_tproxy_setup forks `nft`/`ip` via PATH while privileged. Pin a
  known-safe PATH ("/usr/sbin:/usr/bin:/sbin:/bin") in the child before execvp
  so a polluted service environment cannot substitute an attacker-controlled
  `nft`/`ip`. Portable across Debian/split-usr layouts (vs hardcoding an
  absolute path). Verified e2e still installs rules through the pinned PATH.

- MEDIUM: --tproxy-fwmark / --tproxy-uid (and [Ingress] Fwmark/SkipUid) accepted
  up to LONG_MAX but are narrowed to int/uid_t at apply, silently wrapping
  (wrong-UID exemption or negative mark). Bound all to INT_MAX (CLI + config);
  SkipUid still reserves (uid_t)-1 as the internal "use geteuid()" sentinel.

- NIT: mq_gw_fetch_adapter.h LIFETIME comment listed a teardown order that
  contradicted the sanctioned CLI order (core/gw_client must be freed FIRST,
  while the H3 engine is live, so a late conn-close cb is a no-op). Corrected;
  confirmed mq_gw_fetch_adapter_free does not deref the core (just free(a)).

- NIT: e2e_tproxy.sh comments still described a hardcoded :443 origin; the port
  is dynamic now (free ephemeral + --tproxy-dport). Updated to ${ORIGIN_PORT}.

NOT changed (intentional, per commit 418de6b): best-effort nft install with
idempotent teardown cleanup — codex proposed transactional rollback; kept as the
deliberate design.

Verified: build clean, test_config + test_cli_help pass, e2e_tproxy 3/3 green.

+++ src/config/mq_config.c
@@ -257,7 +257,8 @@ handle_kv(mq_file_config_t *cfg, int section, const char *key, const char *val,
         else if (strcasecmp(key, "Mode") == 0)
             CSTR(tproxy_mode);
         else if (strcasecmp(key, "Fwmark") == 0)
-            LONGV(tproxy_fwmark, 1, LONG_MAX);
+            /* capped at INT_MAX: applied as int (mq_tproxy_setup_t.fwmark). */
+            LONGV(tproxy_fwmark, 1, INT_MAX);
         else if (strcasecmp(key, "Table") == 0)
             LONGV(tproxy_table, 1, 65535);
         else if (strcasecmp(key, "Dport") == 0)
@@ -265,7 +266,9 @@ handle_kv(mq_file_config_t *cfg, int section, const char *key, const char *val,
         else if (strcasecmp(key, "SetupRedirect") == 0)
             cfg->setup_redirect = parse_bool(val);
         else if (strcasecmp(key, "SkipUid") == 0)
-            LONGV(tproxy_skip_uid, -1, LONG_MAX);
+            /* capped at INT_MAX: narrowed to uid_t at apply; reserves (uid_t)-1
+             * as the internal "use geteuid()" sentinel, not a user value. */
+            LONGV(tproxy_skip_uid, -1, INT_MAX);
         else
             MQ_LOGW("%s:%d: unknown key '%s' in [Ingress]", path, lineno, key);
         break;
